VERY HAPPY WITH PROJECT SO FAR!
so I’ve added crap tonnes of methods that redirect from the scene to the skview to the view controller BUT it works! and its performing very well. doesn’t look great yet but at least now theres some sort of creation aspect to the drag element picker.

no known bugs but I imagine there might be 1 or 2 combinations of elements that don’t have names because of alphabetising or something. maybe redundant methods too